Newest version, more robust
Remember to download all files in the same folder. Run the "run_one_time.py" only one time as it will create a unique system id in a txt file on the same folder. This version appends the images to the bottom of the text which I don't think is the way we want to do it in the end. We are working on a new version:)